requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
attrs==25.4.0
bitarray==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.3
ckzg==2.1.5
click==8.3.0
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.118.0
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
multidict==6.7.0
parsimonious==0.10.0
propcache==0.4.0
pycryptodome==3.23.0
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
python-dotenv==1.1.1
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
toolz==1.0.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
web3==7.13.0
websockets==15.0.1
yarl==1.22.0
